Refactor factory relationship handling per past decisions
During the import process, I started using `wp_generate_uuid4()`
as a way to track unique ID. Somehow I ignored that when writing
the code to track relationships after the initial import, so stuff
got super screwy.

This takes the position that a unique ID is generated only in that
way and then relationships should be maintained based on that ID.

This also enables the "sit" data that is already stored in the DB
thanks to the import.